Move HTTPS tests out of version tests section
